refactored processing code
