+++ src/scores/__init__.py
@@ -1,6 +1,7 @@
 """
 The philosphy is to import the public API during the init phase rather than leaving it to the user
 """
+
 # pylint: disable=E0603
 
 import scores.categorical
@@ -8,6 +9,7 @@ import scores.continuous
 import scores.functions
 import scores.pandas
 import scores.probability
+import scores.processing
 import scores.sample_data
 import scores.stats.statistical_tests  # noqa: F401
 
@@ -19,6 +21,7 @@ __all__ = [
     "scores.functions",
     "scores.pandas",
     "scores.probability",
+    "scores.processing",
     "scores.sample_data",
     "scores.stats.statistical_tests",
 ]
