+++ src/scores/processing/__init__.py
@@ -1,0 +1,19 @@
+"""
+Import the functions from the implementations into the public API
+"""
+
+from scores.processing.discretise import (
+    binary_discretise,
+    binary_discretise_proportion,
+    comparative_discretise,
+    proportion_exceeding,
+)
+from scores.processing.matching import broadcast_and_match_nan
+
+__all__ = [
+    "comparative_discretise",
+    "binary_discretise",
+    "proportion_exceeding",
+    "broadcast_and_match_nan",
+    "binary_discretise_proportion",
+]
