+++ tests/processing/test_discretise.py
@@ -1,71 +1,16 @@
-"""Tests for processing"""
-
+"""Tests for scores.processing.discretise"""
 
 import numpy as np
 import pytest
 import xarray as xr
 
 from scores.processing import (
-    _binary_discretise_proportion,
     binary_discretise,
-    broadcast_and_match_nan,
-    check_binary,
     comparative_discretise,
     proportion_exceeding,
 )
-from tests import test_processing_data as xtd
-
-
-@pytest.mark.parametrize(
-    ("args", "expected"),
-    [
-        # DataArrays
-        ([], tuple()),
-        ([xtd.DA_2], (xtd.DA_2,)),
-        ([xtd.DA_1, xtd.DA_2], xtd.EXPECTED_12),
-        ([xtd.DA_2, xtd.DA_1], xtd.EXPECTED_21),
-        ([xtd.DA_1, xtd.DA_2, xtd.DA_3], xtd.EXPECTED_123),
-        ([xtd.DA_2, xtd.DA_4], xtd.EXPECTED_24),
-        ([xtd.DA_1, xtd.DA_5], xtd.EXPECTED_15),
-        ([xtd.DA_2, xtd.DA_6], xtd.EXPECTED_26),
-        ([xtd.DA_3, xtd.DA_7], xtd.EXPECTED_37),
-        # Datasets
-        ([xtd.DS_12], xtd.EXPECTED_DS12),
-        ([xtd.DS_123], xtd.EXPECTED_DS123),
-        ([xtd.DS_12, xtd.DS_123], xtd.EXPECTED_DS12_DS123),
-        ([xtd.DS_12, xtd.DS_3], xtd.EXPECTED_DS12_DS3),
-        # Datasets and DataArrays
-        ([xtd.DS_3, xtd.DA_7], xtd.EXPECTED_DS3_7),
-        ([xtd.DS_7, xtd.DA_3], xtd.EXPECTED_DS7_3),
-        ([xtd.DS_12, xtd.DA_3], xtd.EXPECTED_DS12_3),
-    ],
-)
-def test_broadcast_and_match_nan(args, expected):
-    """
-    Tests that broadcast_and_match_nan calculates the correct result
-    Args:
-        args: a list of the args that will be *-ed into match_DataArray
-        expected: a tuple, the expected output of match_DataArray
-    """
-    calculated = broadcast_and_match_nan(*args)
-    for calculated_element, expected_element in zip(calculated, expected):
-        assert calculated_element.equals(expected_element)
-
-
-@pytest.mark.parametrize(
-    ("args", "error_msg_snippet"),
-    [
-        ([xr.Dataset({"DA_1": xtd.DA_1}), xtd.DA_1, np.arange(4)], "Argument 2"),
-        ([np.arange(5)], "Argument 0"),
-    ],
-)
-def test_broadcast_and_match_nan_rasies(args, error_msg_snippet):
-    """
-    Tests that processing.broadcast_and_match_nan correctly raises an ValueError
-    """
-    with pytest.raises(ValueError) as excinfo:
-        broadcast_and_match_nan(*args)
-    assert error_msg_snippet in str(excinfo.value)
+from scores.processing.discretise import binary_discretise_proportion
+from tests.processing import test_data as xtd
 
 
 @pytest.mark.parametrize(
@@ -424,7 +369,7 @@ def test_binary_discretise(data, thresholds, mode, abs_tolerance, autosqueeze, e
             None,
             False,
             ValueError,
-            "Values in `thresholds` are not montonic increasing",
+            "Values in `thresholds` are not monotonic increasing",
         ),
         # invalid abs_tolerance
         (
@@ -465,37 +410,6 @@ def test_binary_discretise_raises(data, thresholds, mode, abs_tolerance, autosqu
     with pytest.raises(error_class) as exc:
         binary_discretise(data, thresholds, mode, abs_tolerance=abs_tolerance, autosqueeze=autosqueeze)
     assert error_msg_snippet in str(exc.value)
-
-
-@pytest.mark.parametrize(
-    ("da"),
-    [
-        (xr.DataArray([0, 1, 2])),
-        (xr.DataArray([0, 1, -1])),
-        (xr.DataArray([0, 1, 0.5])),
-        (xr.DataArray([[0, 1, 1.0000001], [0, 1, 1]])),
-    ],
-)
-def test_check_binary_raises(da):
-    """test check_binary raises"""
-    with pytest.raises(ValueError) as exc:
-        check_binary(da, "my name")
-    assert "`my name` contains values that are not in the set {0, 1, np.nan}" in str(exc.value)
-
-
-@pytest.mark.parametrize(
-    ("da"),
-    [
-        (xr.DataArray([0, 1])),
-        (xr.DataArray([0, 0])),
-        (xr.DataArray([1, 1])),
-        (xr.DataArray([0, 1, np.nan])),
-        (xr.DataArray([[0, 1, np.nan], [0, 1, np.nan]])),
-    ],
-)
-def test_check_binary_doesnt_raise(da):
-    """test check_binary doesn't raise"""
-    check_binary(da, "my name")
 
 
 @pytest.mark.parametrize(
@@ -603,13 +517,13 @@ def test_proportion_exceeding(data, thresholds, reduce_dims, preserve_dims, expe
         ),
     ],
 )
-def test__binary_discretise_proportion(
+def test_binary_discretise_proportion(
     data, thresholds, mode, reduce_dims, preserve_dims, abs_tolerance, autosqueeze, expected
 ):
     """
-    Tests that processing._binary_discretise_proportion returns the correct value.
+    Tests that processing.binary_discretise_proportion returns the correct value.
     """
-    calc = _binary_discretise_proportion(
+    calc = binary_discretise_proportion(
         data,
         thresholds,
         mode,
